no practice method

+++ src/events/events.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   Logger,
@@ -26,7 +27,7 @@ import { EventsService } from './events.service';
 import { ListEvents } from './dtos/list.events';
 import { CurrentUser } from '../auth/current-user.decorator';
 import { User } from '../auth/user.entity';
-import { AuthGuardJwt } from "../auth/auth-guard-jwt";
+import { AuthGuardJwt } from '../auth/auth-guard-jwt';
 
 @Controller({
   path: '/events',
@@ -59,35 +60,35 @@ export class EventsController {
     return events;
   }
 
-  @Get('practice2')
-  async practice2() {
-    // return await this.eventRepository.findOne({
-    //   where: {
-    //     id: 1,
-    //   },
-    //   // loadEagerRelations: false
-    //   relations: ['attendees'],
-    // });
+  // @Get('practice2')
+  // async practice2() {
+  //   // return await this.eventRepository.findOne({
+  //   //   where: {
+  //   //     id: 1,
+  //   //   },
+  //   //   // loadEagerRelations: false
+  //   //   relations: ['attendees'],
+  //   // });
 
-    // const event = await this.eventRepository.findOne({
-    //   where: {
-    //     id: 1,
-    //   },
-    // });
+  //   // const event = await this.eventRepository.findOne({
+  //   //   where: {
+  //   //     id: 1,
+  //   //   },
+  //   // });
 
-    // const attendee = new Attendee();
-    // attendee.name = 'jerry';
-    // attendee.event = event;
-    // await this.attendeeRepository.save(attendee);
-    // return event;
+  //   // const attendee = new Attendee();
+  //   // attendee.name = 'jerry';
+  //   // attendee.event = event;
+  //   // await this.attendeeRepository.save(attendee);
+  //   // return event;
 
-    return this.eventRepository
-      .createQueryBuilder('e')
-      .select(['e.id', 'e.name'])
-      .orderBy('e.id', 'ASC')
-      .take(3)
-      .getMany();
-  }
+  //   // return this.eventRepository
+  //   //   .createQueryBuilder('e')
+  //   //   .select(['e.id', 'e.name'])
+  //   //   .orderBy('e.id', 'ASC')
+  //   //   .take(3)
+  //   //   .getMany();
+  // }
 
   @Get(':id')
   async findOne(@Param('id', new ParseIntPipe()) id: number) {
@@ -104,12 +105,38 @@ export class EventsController {
   async create(@Body() input: CreateEventDto, @CurrentUser() user: User) {
     return await this.eventsService.createEvent(input, user);
   }
+  
 
   @Patch(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(200)
   async update(
     @Param('id', new ParseIntPipe()) id: number,
     @Body() input: UpdateEventDto,
+    @CurrentUser() user: User,
+  ) {
+    const event = await this.eventsService.getEvent(id);
+
+    if(!event){
+      throw new NotFoundException();
+    }
+
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to change this event`,
+      );
+    }
+
+    return await this.eventsService.updateEvent(event, input);
+  }
+
+  @Delete(':id')
+  @UseGuards(AuthGuardJwt)
+  @HttpCode(204)
+  async remove(
+    @Param('id', new ParseIntPipe()) id: number,
+    @CurrentUser() user: User,
   ) {
     const event = await this.eventRepository.findOne({
       where: {
@@ -117,21 +144,17 @@ export class EventsController {
       },
     });
 
-    const updatedEvent = {
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when,
-    };
-
-    return this.eventRepository.save(updatedEvent);
-  }
-
-  @Delete(':id')
-  @HttpCode(204)
-  async remove(@Param('id', new ParseIntPipe()) id: number) {
-    const result = await this.eventsService.deleteEvent(id);
-    if (result?.affected !== 1) {
+    if(!event){
       throw new NotFoundException();
     }
+
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to change this event`,
+      );
+    }
+
+    await this.eventsService.deleteEvent(id);
   }
 }
